final commit

+++ src/SalesManager.java
@@ -14,8 +14,8 @@ public class SalesManager {
         }
         return max;
     }
-    public int min() {
-        int min = sales[0];
+    public long min() {
+        long min = sales[0];
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] < min) {
                 min = sales[i];
@@ -23,8 +23,8 @@ public class SalesManager {
         }
         return min;
     }
-    public int average(){
-        int average=0;
+    public long average(){
+        long average=0;
         for (int i = 0; i < sales.length; i++) {
             average+=sales[i];
         }
